# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cdd-ctl",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/auth/github": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login with GitHub OAuth",
        "operationId": "login_github",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully authenticated via GitHub",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "OAuth failed"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login with username/password",
        "operationId": "login_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register a new user",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthResponse": {
        "type": "object",
        "description": "Response containing the JWT token",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT token"
          }
        }
      },
...
      "LoginPayload": {
        "type": "object",
        "description": "Payload for logging in with username/password",
        "required": [
          "username"
        ],
        "properties": {
          "password": {
            "type": [
              "string",
              "null"
            ],
            "description": "Password"
          },
          "username": {
            "type": "string",
            "description": "Username"
          }
        }
      },
      "OAuthPayload": {
        "type": "object",
        "description": "Payload for OAuth login",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "OAuth authorization code"
          }
        }
      },
...
      "RegisterPayload": {
        "type": "object",
        "description": "Payload for registering a new user",
        "required": [
          "username",
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address"
          },
          "password": {
            "type": [
              "string",
              "null"
            ],
            "description": "Password"
          },
          "username": {
            "type": "string",
            "description": "Desired username"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}